Added the ControlSequenceHandler for matching debug and cheat sequence as well as user-defined touch and key sequences.
This class will track touches/taps on the screen as well as presses on
a number keypad. The screen is divided logically to form a 3x3 grid
representing the digit keys of a convential mobile phone, 1..9.

If the user taps a certain sequence, a debug or cheat event will be
triggered through the SystemContext methods.

Additionaly the isSequenceMatched can be called with an arbitrary
sequence to trigger application-specific events.